Homework3 fix var type

+++ Homework1/src/Homework3/task_2/Employee.java
@@ -5,10 +5,10 @@ public class Employee {
     private final String position;
     private final String email;
     private final String phone;
-    private double salary;
-    private short age;
+    private int salary;
+    private int age;
 
-    public Employee(String fullName, String position, String email, String phone, double salary, short age) {
+    public Employee(String fullName, String position, String email, String phone, int salary, int age) {
         this.fullName = fullName;
         this.position = position;
         this.email = email;
